moving around classes so that they are better organized
