+++ java/src/main/java/com/saucelabs/simplesauce/ConcreteSystemManager.java
@@ -1,5 +1,7 @@
 package com.saucelabs.simplesauce;
 
+import com.saucelabs.simplesauce.interfaces.EnvironmentManager;
+
 public class ConcreteSystemManager implements EnvironmentManager {
     public String getEnvironmentVariable(String variable) {
         return System.getenv(variable);
