+++ java/src/main/java/com/saucelabs/simplesauce/DataCenter.java
@@ -4,4 +4,6 @@ public enum DataCenter
 {
     ;
     public static String USWest = "https://ondemand.saucelabs.com/wd/hub";
+    public static final String US_EAST  = "https://ondemand.us-east-1.saucelabs.com/wd/hub";
+    public static final String EU_CENTRAL = "https://ondemand.eu-central-1.saucelabs.com/wd/hub";
 }
