--- java/src/main/java/com/saucelabs/simplesauce/config/SauceURL.java
@@ -1,8 +1,0 @@
-package com.saucelabs.simplesauce.config;
-
-public class SauceURL
-{
-	public static final String US_WEST = "https://ondemand.saucelabs.com/wd/hub";
-	public static final String US_EAST  = "https://ondemand.us-east-1.saucelabs.com/wd/hub";
-	public static final String EU_CENTRAL = "https://ondemand.eu-central-1.saucelabs.com/wd/hub";
-}
