+++ java/src/main/java/com/saucelabs/simplesauce/interfaces/EnvironmentManager.java
@@ -1,4 +1,4 @@
-package com.saucelabs.simplesauce;
+package com.saucelabs.simplesauce.interfaces;
 
 public interface EnvironmentManager {
     String getEnvironmentVariable(String variable);
